requirements.txt
absl-py==2.1.0
accelerate==0.30.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autoflake==2.2.0
av==12.0.0
azure-core==1.30.1
azure-identity==1.15.0
backoff==2.2.1
basicsr==1.4.2
bcrypt==4.1.2
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
black==22.12.0
blake3==0.3.4
boto3==1.34.138
botocore==1.34.138
Brotli==1.0.9
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
conda==4.12.0
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.4
datasets==2.18.0
Deprecated==1.2.14
diffusers==0.27.2
dill==0.3.8
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
docker==7.0.0
einops==0.8.0
exceptiongroup==1.2.0
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-generativeai==0.4.1
google-resumable-media==2.7.1
googleapis-common-protos==1.63.0
gradio==4.24.0
gradio_client==0.14.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
imageio==2.34.1
importlib-metadata==7.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
inquirerpy==0.3.4
iopath==0.1.10
isort==5.13.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json_repair==0.10.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kubernetes==29.0.0
lazy_loader==0.4
libcst==1.1.0
lightning-utilities==0.11.2
litellm==1.34.18
llama-index==0.10.26
llama-index-agent-openai==0.2.1
llama-index-cli==0.1.11
llama-index-core==0.10.26
llama-index-embeddings-azure-openai==0.1.6
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-ollama==0.1.2
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-azure-openai==0.1.5
llama-index-llms-openai==0.1.14
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.13
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-chroma==0.1.6
llama-parse==0.4.0
llamaindex-py-client==0.1.15
lmdb==1.4.1
loguru==0.7.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matrix-client==0.4.0
mdurl==0.1.2
minijinja==1.0.16
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
msgpack==1.0.8
msgpack-numpy==0.4.8
multidict==6.0.5
multiprocess==0.70.16
mypy==1.9.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.16.2
openai==1.14.3
opencv-python==4.9.0.80
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.0
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
pfzy==0.3.4
pillow==10.3.0
pip==23.3
pipenv==2023.12.1
platformdirs==4.2.0
pluggy==1.5.0
portalocker==2.8.2
posthog==3.5.0
pretty-errors==1.2.25
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pulsar-client==3.4.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycosat==0.6.6
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
pyflakes==3.2.0
PyGithub==2.3.0
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.2
pypdf==4.1.0
PyPika==0.48.9
pyproject_hooks==1.0.0
PySocks==1.7.1
pytest==8.2.2
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-on-whales==0.68.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rich-click==1.8.3
rpds-py==0.18.1
rsa==4.9
ruamel-yaml-conda==0.15.100
ruff==0.3.4
s3transfer==0.10.2
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==2.6.1
setuptools==69.2.0
shellingham==1.5.4
single-source==0.3.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.36.3
striprtf==0.0.26
sympy==1.12
tabulate==0.9.0
tb-nightly==2.17.0a20240507
tenacity==8.2.3
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2024.5.3
tiktoken==0.6.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.0
torchaudio==2.3.0
torchmetrics==1.4.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.40.2
truss==0.9.18
typer==0.12.0
typer-cli==0.12.0
typer-slim==0.12.0
typing_extensions==4.10.0
typing-inspect==0.9.0
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.25.1
watchfiles==0.19.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.3
wheel==0.41.2
wrapt==1.16.0
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
zstandard==0.19.0
